+++ src/utils/checkScreenSize/checkScreenSize.tsx
@@ -1,0 +1,5 @@
+export const checkScreenSize = () => {
+    return (window.innerWidth < 871);
+};
+
+console.log(checkScreenSize)
